code/requirements.minimal.txt
# --- Core Dependencies for Text2SQL Project ---

# Database & Data Generation
pandas
Faker
SQLAlchemy

# Milvus/Zilliz Cloud & Embeddings
pymilvus
pymilvus-model
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu  # 强制安装CPU版本

# LLM & Environment
langchain-deepseek
python-dotenv